WIP tests for reverse
